meta-auto-country-feed/requirements.txt
python-dotenv==1.1.1
ShopifyAPI==12.7.0
google-api-python-client==2.176.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
